__archives_iac/workers/loader/requirements.txt
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.14.0
pandas>=1.5.0
pyarrow>=14.0.0
db-dtypes>=1.2.0
